requirements.txt
flask
flask-socketio
pandas
scikit-learn==1.6.1
numpy
joblib
kafka-python
mysql-connector-python
requests
python-dateutil
gunicorn
eventlet
xgboost 